move BART to its own module
