+++ pymc/tests/test_bart.py
@@ -7,7 +7,7 @@ import pymc as pm
 
 
 def test_split_node():
-    split_node = pm.distributions.tree.SplitNode(index=5, idx_split_variable=2, split_value=3.0)
+    split_node = pm.bart.tree.SplitNode(index=5, idx_split_variable=2, split_value=3.0)
     assert split_node.index == 5
     assert split_node.idx_split_variable == 2
     assert split_node.split_value == 3.0
@@ -18,7 +18,7 @@ def test_split_node():
 
 
 def test_leaf_node():
-    leaf_node = pm.distributions.tree.LeafNode(index=5, value=3.14, idx_data_points=[1, 2, 3])
+    leaf_node = pm.bart.tree.LeafNode(index=5, value=3.14, idx_data_points=[1, 2, 3])
     assert leaf_node.index == 5
     assert np.array_equal(leaf_node.idx_data_points, [1, 2, 3])
     assert leaf_node.value == 3.14
